Add a rough measure for how many scripts were fixed across databases.
This compares the snippets in scratch and live database to see how many are
gone, and cross-references those with the ones that had been clicked.

This is of course prone to false positives (snippet got moved, reworded, or
simply fixed by someone else not through CH), but it's a start.


Former-commit-id: 61a1726f14f2b2058d730bec9453cdb8bde4f691