Revert database connection pooling stuff.
This is likely way more trouble than it's worth - it makes the code a
bit simple outside of chdb.py, but introduces some complexity to chdb.py
that can be hard to debug.

Furthermore, this only really saves about 10ms per request, so it's a
very premature optimization at this point.