When extracting sections, throw away the citation needed templates.
dewiki has started wrapping our <span class="ch-cn-marker"> in <p>
when converting to HTML, but because we already removed the contents
of the template implicitly doing cleanup, we ended up with:

<p><span class="ch-cn-marker"></span></p>

in the snippet, which broke the UI because it assumes the marker is
preceded by some text content.

We now don't bother keeping the markers around. This does mean we
couldn't extract full sections in a wiki with inline {{cn}}, so
extracting snippets and sections are now further apart from one another.